refactor code to match new Program module
