+++ zkvm/src/lib.rs
@@ -13,6 +13,7 @@ mod merkle;
 mod ops;
 mod point_ops;
 mod predicate;
+mod program;
 mod prover;
 mod scalar_witness;
 mod signature;
@@ -26,8 +27,9 @@ pub use self::constraints::{Commitment, Constraint, Expression, Variable};
 pub use self::contract::{Anchor, Contract, ContractID, Output, PortableItem};
 pub use self::errors::VMError;
 pub use self::merkle::{MerkleItem, MerkleNeighbor, MerkleTree};
-pub use self::ops::{Instruction, Opcode, PredicateTree, Program};
+pub use self::ops::{Instruction, Opcode};
 pub use self::predicate::Predicate;
+pub use self::program::Program;
 pub use self::prover::Prover;
 pub use self::scalar_witness::ScalarWitness;
 pub use self::signature::{Signature, VerificationKey};
